chore(web): move mixPanel tracing provider to the observability/providers folder
