+++ apps/web/src/pages/_app.tsx
@@ -91,7 +91,7 @@ import useInvalidateOverviewsOnTx from '@/hooks/useInvalidateOverviewsOnTx'
 import { GATEWAY_URL } from '@/config/gateway'
 import { captureError, initObservability } from '@/services/observability'
 import { DatadogProvider } from '@/services/observability/providers/datadog'
-import { MixpanelTracingProvider } from '@/services/analytics/MixpanelTracingProvider'
+import { MixpanelTracingProvider } from '@/services/observability/providers/mixpanel'
 import useMixpanel from '@/services/analytics/useMixpanel'
 import { AddressBookSourceProvider } from '@/components/common/AddressBookSourceProvider'
 import { CaptchaProvider } from '@/components/common/Captcha'
