+++ apps/web/src/services/observability/providers/__tests__/mixpanel.test.ts
@@ -1,7 +1,7 @@
-import { MixpanelTracingProvider } from '../MixpanelTracingProvider'
-import { trackErrorSurfaced } from '../error-tracking'
+import { MixpanelTracingProvider } from '../mixpanel'
+import { trackErrorSurfaced } from '@/services/analytics/error-tracking'
 
-jest.mock('../error-tracking', () => ({
+jest.mock('@/services/analytics/error-tracking', () => ({
   trackErrorSurfaced: jest.fn(),
 }))
